refactored code (commited using npm script)

+++ tests/blog_api.test.js
@@ -4,67 +4,18 @@ const mongoose = require('mongoose')
 const supertest = require('supertest')
 const app = require('../app')
 const Blog = require('../models/blog')
+const helper = require('./helper')
+const User = require('../models/user')
+const bcrypt = require('bcrypt')
+const logger = require('../utils/logger')
 
 const api = supertest(app)
 
 beforeEach(async () => {
-  await Blog.deleteMany({})
-
-  const blogObjects = initialBlogs.map(blog => new Blog(blog))
-  const promiseArray = blogObjects.map(blog => blog.save())
-  await Promise.all(promiseArray)
+  await helper.resetUsers()
+  await helper.resetBlogs()
 })
 
-const initialBlogs = [
-  {
-    _id: "5a422a851b54a676234d17f7",
-    title: "React patterns",
-    author: "Michael Chan",
-    url: "https://reactpatterns.com/",
-    likes: 7,
-    __v: 0
-  },
-  {
-    _id: "5a422aa71b54a676234d17f8",
-    title: "Go To Statement Considered Harmful",
-    author: "Edsger W. Dijkstra",
-    url: "http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html",
-    likes: 5,
-    __v: 0
-  },
-  {
-    _id: "5a422b3a1b54a676234d17f9",
-    title: "Canonical string reduction",
-    author: "Edsger W. Dijkstra",
-    url: "http://www.cs.utexas.edu/~EWD/transcriptions/EWD08xx/EWD808.html",
-    likes: 12,
-    __v: 0
-  },
-  {
-    _id: "5a422b891b54a676234d17fa",
-    title: "First class tests",
-    author: "Robert C. Martin",
-    url: "http://blog.cleancoder.com/uncle-bob/2017/05/05/TestDefinitions.htmll",
-    likes: 10,
-    __v: 0
-  },
-  {
-    _id: "5a422ba71b54a676234d17fb",
-    title: "TDD harms architecture",
-    author: "Robert C. Martin",
-    url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
-    likes: 0,
-    __v: 0
-  },
-  {
-    _id: "5a422bc61b54a676234d17fc",
-    title: "Type wars",
-    author: "Robert C. Martin",
-    url: "http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html",
-    likes: 2,
-    __v: 0
-  }
-]
 
 
 // DO NOT RUN ALL TESTS AT ONCE. SOME MIGHT FAIL DUE TO ASYNCRONOUS FUNCTIONALITY.
@@ -81,7 +32,7 @@ describe('GET /api/json tests', () => {
   
   test('correct amount of blogs is returned', async () => {
     const response = await api.get('/api/blogs')
-    assert.strictEqual(response.body.length, initialBlogs.length)
+    assert.strictEqual(response.body.length, helper.initialBlogs.length)
   })
   
   test('id property of the blog posts is defined', async () => {
@@ -91,89 +42,76 @@ describe('GET /api/json tests', () => {
 })
 
 describe('POST /api/json tests', () => {
-  test('a valid blog can be added using the post method', async () => {
+  test('a valid blog can be added using the post method and a user auth token', async () => {
     const initialBlogs = await api.get('/api/blogs')
-  
-    const newBlog = {
-      title: "This is a valid blog!",
-      author: "Jane Doe",
-      url: "http://trial.com/",
-      likes: 4
-    }
-  
+
+    const user = await User.findOne({}) 
+    const token = helper.generateUserToken(user)
+
     await api
       .post('/api/blogs')
-      .send(newBlog)
-  
+      .set('Authorization', `Bearer ${token}`)
+      .send(helper.sampleBlog)
+
     const response = await api.get('/api/blogs')
     assert(response.body.length === initialBlogs.body.length + 1)
   })
   
   test('if likes property is missing from added blog, its value defaults to 0 and is added', async () => {
-    const initialBlogs = await api.get('/api/blogs')
-  
-    const newBlog = {
-      title: "This does not have a likes property!",
-      author: "Jane Doe",
-      url: "http://no_likes.com/"
-    }
-  
+    const user = await User.findOne({}) 
+    const token = helper.generateUserToken(user)
+
     const response = await api
       .post('/api/blogs')
-      .send(newBlog)
-  
+      .set('Authorization', `Bearer ${token}`)
+      .send({...helper.sampleBlog, likes: null})
+    
     assert(response.body.likes === 0)
   })
   
   test('if title property is missing from added blog, backend replies with status code 400', async () => {
     const initialBlogs = await api.get('/api/blogs')
-    const newBlog = {
-      author: "Jane Doe",
-      url: "http://no_title.com/"
-    }
-  
+
+    const user = await User.findOne({}) 
+    const token = helper.generateUserToken(user)
+
     await api
       .post('/api/blogs')
-      .send(newBlog)
+      .set('Authorization', `Bearer ${token}`)
+      .send({...helper.sampleBlog, title: null})
       .expect(400)
   
-    const response = await api.get('/api/notes')
-  
+      const response = await api.get('/api/notes')
     assert.strictEqual(response.body.length, initialBlogs.length)
   })
   
-  test('if url property is missing from added blog, backend replies with status code 400', async () => {
+  test('quirky if url property is missing from added blog, backend replies with status code 400', async () => {
     const initialBlogs = await api.get('/api/blogs')
-    const newBlog = {
-      author: "Jane Doe",
-      title: "There is no URL!"
-    }
-  
+
+    const user = await User.findOne({}) 
+    const token = helper.generateUserToken(user)
+
     await api
       .post('/api/blogs')
-      .send(newBlog)
+      .set('Authorization', `Bearer ${token}`)
+      .send({...helper.sampleBlog, url: null})
       .expect(400)
   
-    const response = await api.get('/api/notes')
-  
+      const response = await api.get('/api/notes')
     assert.strictEqual(response.body.length, initialBlogs.length)
   })
 })
 
 describe('DELETE /api/json/:id tests', () => {
-  beforeEach(async () => {
-    await Blog.deleteMany({})
-  
-    const blogObjects = initialBlogs.map(blog => new Blog(blog))
-    const promiseArray = blogObjects.map(blog => blog.save())
-    await Promise.all(promiseArray)
-  })
 
   test('a blog can be deleted from the backend', async () => { 
     const initialBlogs = await api.get('/api/blogs')
     const blogToDelete = initialBlogs.body[0]
 
-    await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204)
+    const user = await User.findOne({username: "sample"}) 
+    const token = helper.generateUserToken(user)
+
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set('Authorization', `Bearer ${token}`).expect(204)
     const laterBlogs = await api.get('/api/blogs')
     
     await api.get(`/api/blogs/${blogToDelete.id}`).expect(404)
@@ -186,6 +124,9 @@ describe("PUT /api/blogs/:id", () => {
     const initialBlogs = await api.get('/api/blogs')
     const blogToUpdate = initialBlogs.body[0]
 
+    const user = await User.findOne({username: "sample"}) 
+    const token = helper.generateUserToken(user)
+
     const updatedBlog = {
       title: `${blogToUpdate.title} (updated)`,
       author: `${blogToUpdate.author} (updated)`,
@@ -195,7 +136,7 @@ describe("PUT /api/blogs/:id", () => {
 
     await api
       .put(`/api/blogs/${blogToUpdate.id}`)
-      .send(updatedBlog)
+      .send(updatedBlog).set('Authorization', `Bearer ${token}`)
       .expect(200)
 
     const response = await api.get(`/api/blogs/${blogToUpdate.id}`)
